requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
chromadb
faiss-cpu
langchain_google_genai
gunicorn
langchain_text_splitters
langchain_chroma